once again new heuristics

+++ Heuristics.py
@@ -57,8 +57,8 @@ class Heuristics:
 
     @staticmethod
     def sum_shortest_distance_and_euclid(node, player_pos, end_tile_pos, weight_euclid=1, weight_path=1):
-        return Heuristics.euclid(player_pos, end_tile_pos) * weight_euclid + \
-               Heuristics.shortest_distance_end_path_player_path(node, player_pos, end_tile_pos) * weight_path
+        return int(Heuristics.euclid(player_pos, end_tile_pos) * weight_euclid + \
+               Heuristics.shortest_distance_end_path_player_path(node, player_pos, end_tile_pos) * weight_path)
 
 
 class HeuristicHelpers:
